+++ example/lib/main.dart
@@ -1,6 +1,8 @@
 import 'package:example/data.dart';
 import 'package:example/multi_select_field_examples/multi_select_field_demo.dart';
 import 'package:example/multi_select_field_examples/multi_select_options_control.dart';
+import 'package:example/multi_select_field_examples/search_multi_select_field.dart';
+import 'package:example/select_field_examples/search_select_field.dart';
 import 'package:example/select_field_examples/custom_menu_select_field.dart';
 import 'package:example/select_field_examples/styled_select_field.dart';
 import 'package:flutter/material.dart';
@@ -40,7 +42,7 @@ class _MyHomePageState extends State<MyHomePage> {
     return Scaffold(
       appBar: AppBar(
         title: const Text(
-          'Multi Select Field',
+          'Search and Select',
           style: TextStyle(
             color: Colors.white,
             fontWeight: FontWeight.bold,
@@ -59,7 +61,11 @@ class _MyHomePageState extends State<MyHomePage> {
             const SizedBox(height: 32),
             MultiSelectOptionsControl<String>(options: options),
             const SizedBox(height: 32),
-            MultiSelectFieldDemo<String>(options: options)
+            MultiSelectFieldDemo<String>(options: options),
+            const SizedBox(height: 32),
+            SearchSelectField(options: options),
+            const SizedBox(height: 32),
+            SearchMultiSelectField(options: options),
           ],
         ),
       ),