+++ lib/select_field.dart
@@ -6,3 +6,4 @@ export 'src/menu_decoration.dart';
 export 'src/select_field/select_field_menu_controller.dart';
 export 'src/multi_select_field/multi_select_field.dart';
 export 'src/multi_select_field/multi_select_field_menu_controller.dart';
+export 'src/search_options.dart';