+++ lib/src/multi_select_field/multi_select_field.dart
@@ -17,10 +17,14 @@ class MultiSelectField<T> extends StatefulWidget {
   /// Callback for providing selected options list
   final void Function(List<Option<T>> options)? onOptionsSelected;
 
-  /// Allows controlling menu's behavior and provides selected options for custom usage
+  /// Allows controlling menu's behavior and provides selected options for custom usage.
+  ///
+  /// By providing [MultiSelectFieldMenuController] you are responsible for `menu's` behaviour.
   final MultiSelectFieldMenuController<T>? menuController;
 
   /// Specifies the [TextEditingController] for [SelectField].
+  ///
+  /// By providing [TextEditingController] you are responsible for `text` behaviour.
   final TextEditingController? textController;
 
   /// See [FocusNode]
@@ -112,6 +116,9 @@ class MultiSelectField<T> extends StatefulWidget {
     void Function(Option<T> option) onOptionSelected,
   )? optionBuilder;
 
+  /// Search is enabled by providing search options. Note that options height is now fixed to a provided value.
+  final SearchOptions<T>? searchOptions;
+
   /// Restoration ID to save and restore the state of the form field.
   final String? restorationId;
 
@@ -145,6 +152,7 @@ class MultiSelectField<T> extends StatefulWidget {
     this.fieldText,
     this.onTextChanged,
     this.onOptionsSelected,
+    this.menuController,
     this.textController,
     this.focusNode,
     this.onTap,
@@ -159,6 +167,7 @@ class MultiSelectField<T> extends StatefulWidget {
     this.suffixIconBuilder,
     this.prefixIconBuilder,
     this.optionBuilder,
+    this.searchOptions,
     this.restorationId,
     this.strutStyle,
     this.textDirection,
@@ -167,7 +176,6 @@ class MultiSelectField<T> extends StatefulWidget {
     this.enabled,
     this.onSaved,
     this.validator,
-    this.menuController,
   });
 
   @override
@@ -178,9 +186,14 @@ class _MultiSelectFieldState<T> extends State<MultiSelectField<T>> {
   late final TextEditingController textController;
   late final MultiSelectFieldMenuController<T> menuController;
   late final FocusNode focusNode;
+  late final bool searchEnabled;
+  late final bool isMenuControllerProvided;
+  late final bool isTextControllerProvided;
 
-  void setControllerText() {
-    if (widget.fieldText != null) {
+  void setControllerText([String? value]) {
+    if (value != null) {
+      textController.text = value;
+    } else if (widget.fieldText != null) {
       textController.text = widget.fieldText!;
     } else {
       final options =
@@ -200,14 +213,23 @@ class _MultiSelectFieldState<T> extends State<MultiSelectField<T>> {
       widget.onOptionsSelected!(selectedOptions);
     }
     menuController.selectedOptions = selectedOptions;
-    setControllerText();
+    if (!searchEnabled && !isTextControllerProvided) {
+      setControllerText();
+    }
   }
 
   void handleOnTap() async {
     if (widget.onTap != null) {
       widget.onTap!();
     }
-    menuController.isExpanded = !menuController.isExpanded;
+    if (searchEnabled &&
+        !menuController.isExpanded &&
+        !isTextControllerProvided) {
+      setControllerText('');
+    }
+    if (!isMenuControllerProvided) {
+      menuController.isExpanded = !menuController.isExpanded;
+    }
   }
 
   void handleOnTapOutside() async {
@@ -215,8 +237,10 @@ class _MultiSelectFieldState<T> extends State<MultiSelectField<T>> {
       widget.onTapOutside!();
     }
 
-    if (menuController.isExpanded) {
-      menuController.isExpanded = false;
+    if (!isMenuControllerProvided) {
+      if (menuController.isExpanded) {
+        menuController.isExpanded = false;
+      }
     }
   }
 
@@ -227,10 +251,15 @@ class _MultiSelectFieldState<T> extends State<MultiSelectField<T>> {
     textController =
         widget.textController ?? TextEditingController(text: widget.fieldText);
     menuController = widget.menuController ?? MultiSelectFieldMenuController();
+    isMenuControllerProvided = widget.menuController != null;
+    isTextControllerProvided = widget.textController != null;
+    searchEnabled = widget.searchOptions != null;
 
     if (widget.initialOptions != null) {
       menuController.selectedOptions = widget.initialOptions!;
-      setControllerText();
+      if (!searchEnabled) {
+        setControllerText();
+      }
     }
 
     menuController.addListener(() async {
@@ -238,8 +267,25 @@ class _MultiSelectFieldState<T> extends State<MultiSelectField<T>> {
         await Future.delayed(widget.menuDecoration?.animationDuration ??
             const Duration(milliseconds: 350));
         focusNode.unfocus();
+
+        if (searchEnabled &&
+            widget.fieldText != null &&
+            !isTextControllerProvided) {
+          setControllerText();
+        }
       }
     });
+
+    if (widget.searchOptions != null) {
+      focusNode.addListener(() {
+        if (!focusNode.hasFocus && menuController.isExpanded) {
+          menuController.isExpanded = false;
+          if (!isTextControllerProvided) {
+            setControllerText();
+          }
+        }
+      });
+    }
   }
 
   @override
@@ -270,6 +316,7 @@ class _MultiSelectFieldState<T> extends State<MultiSelectField<T>> {
       focusNode: focusNode,
       onTap: handleOnTap,
       inputDecoration: widget.inputDecoration,
+      searchOptions: widget.searchOptions,
       restorationId: widget.restorationId,
       inputStyle: widget.inputStyle,
       textAlign: widget.textAlign,